Replace vec complete functions with df complete equivalents
