+++ NAMESPACE
@@ -441,8 +441,11 @@ export(allow_lossy_cast)
 export(as_list_of)
 export(data_frame)
 export(df_cast)
+export(df_detect_complete)
 export(df_list)
+export(df_locate_complete)
 export(df_ptype2)
+export(df_slice_complete)
 export(field)
 export(fields)
 export(is_list_of)
@@ -526,7 +529,6 @@ export(vec_count)
 export(vec_data)
 export(vec_default_cast)
 export(vec_default_ptype2)
-export(vec_detect_complete)
 export(vec_duplicate_any)
 export(vec_duplicate_detect)
 export(vec_duplicate_id)
@@ -544,7 +546,6 @@ export(vec_is)
 export(vec_is_empty)
 export(vec_is_list)
 export(vec_list_cast)
-export(vec_locate_complete)
 export(vec_match)
 export(vec_math)
 export(vec_math_base)
@@ -591,7 +592,6 @@ export(vec_set_names)
 export(vec_size)
 export(vec_size_common)
 export(vec_slice)
-export(vec_slice_complete)
 export(vec_sort)
 export(vec_split)
 export(vec_type)
