+++ src/complete.c
@@ -1,20 +1,21 @@
 #include "vctrs.h"
 #include "equal.h"
+#include "type-data-frame.h"
 
 // -----------------------------------------------------------------------------
 
-static SEXP vec_slice_complete(SEXP x);
+static SEXP df_slice_complete(SEXP x);
 
 // [[ register() ]]
-SEXP vctrs_slice_complete(SEXP x) {
-  return vec_slice_complete(x);
+SEXP vctrs_df_slice_complete(SEXP x) {
+  return df_slice_complete(x);
 }
 
-static SEXP vec_locate_complete(SEXP x);
+static SEXP df_locate_complete(SEXP x);
 
 static
-SEXP vec_slice_complete(SEXP x) {
-  SEXP loc = PROTECT(vec_locate_complete(x));
+SEXP df_slice_complete(SEXP x) {
+  SEXP loc = PROTECT(df_locate_complete(x));
 
   // Skip `vec_as_location()` in `vec_slice()`
   SEXP out = vec_slice_impl(x, loc);
@@ -26,15 +27,15 @@ SEXP vec_slice_complete(SEXP x) {
 // -----------------------------------------------------------------------------
 
 // [[ register() ]]
-SEXP vctrs_locate_complete(SEXP x) {
-  return vec_locate_complete(x);
+SEXP vctrs_df_locate_complete(SEXP x) {
+  return df_locate_complete(x);
 }
 
-static SEXP vec_detect_complete(SEXP x);
+static SEXP df_detect_complete(SEXP x);
 
 static
-SEXP vec_locate_complete(SEXP x) {
-  SEXP where = PROTECT(vec_detect_complete(x));
+SEXP df_locate_complete(SEXP x) {
+  SEXP where = PROTECT(df_detect_complete(x));
   SEXP out = r_lgl_which(where, false);
   UNPROTECT(1);
   return out;
@@ -42,112 +43,49 @@ SEXP vec_locate_complete(SEXP x) {
 
 // -----------------------------------------------------------------------------
 
-static inline void lgl_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void int_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void dbl_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void cpl_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void chr_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void raw_detect_complete(SEXP x, R_len_t size, int* p_out);
-static inline void list_detect_complete(SEXP x, R_len_t size, int* p_out);
-static void df_detect_complete(SEXP x, R_len_t size, int* p_out);
-
 // [[ register() ]]
-SEXP vctrs_detect_complete(SEXP x) {
-  return vec_detect_complete(x);
+SEXP vctrs_df_detect_complete(SEXP x) {
+  return df_detect_complete(x);
 }
 
+static inline void vec_detect_complete_col(SEXP x, R_len_t size, int* p_out, struct df_short_circuit_info* p_info);
+
 static
-SEXP vec_detect_complete(SEXP x) {
-  SEXP proxy = PROTECT(vec_proxy_equal(x));
-  R_len_t size = vec_size(proxy);
+SEXP df_detect_complete(SEXP x) {
+  int nprot = 0;
 
-  SEXP out = PROTECT(Rf_allocVector(LGLSXP, size));
-  int* p_out = LOGICAL(out);
-
-  switch (vec_proxy_typeof(proxy)) {
-  case vctrs_type_logical: lgl_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_integer: int_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_double: dbl_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_complex: cpl_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_character: chr_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_raw: raw_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_list: list_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_dataframe: df_detect_complete(proxy, size, p_out); break;
-  case vctrs_type_scalar: stop_internal("vec_detect_complete", "Can't detect missing values in scalars.");
-  default: stop_unimplemented_vctrs_type("vec_detect_complete", vec_proxy_typeof(proxy));
+  if (!is_data_frame(x)) {
+    r_abort("`x` must be a data frame.");
   }
 
-  UNPROTECT(2);
-  return out;
-}
+  // Flatten df-cols, but don't proxy
+  x = PROTECT_N(df_flatten(x), &nprot);
 
-#define VEC_DETECT_COMPLETE(CTYPE, CONST_DEREF, SCALAR_EQUAL_MISSING) { \
-  const CTYPE* p_x = CONST_DEREF(x);                                    \
-                                                                        \
-  for (R_len_t i = 0; i < size; ++i) {                                  \
-    const CTYPE elt = p_x[i];                                           \
-    p_out[i] = !SCALAR_EQUAL_MISSING(elt);                              \
-  }                                                                     \
-}
+  R_len_t size = vec_size(x);
+  r_ssize n_cols = r_length(x);
 
-static inline
-void lgl_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(int, LOGICAL_RO, lgl_equal_missing_scalar);
-}
-static inline
-void int_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(int, INTEGER_RO, int_equal_missing_scalar);
-}
-static inline
-void dbl_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(double, REAL_RO, dbl_equal_missing_scalar);
-}
-static inline
-void cpl_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(Rcomplex, COMPLEX_RO, cpl_equal_missing_scalar);
-}
-static inline
-void chr_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(SEXP, STRING_PTR_RO, chr_equal_missing_scalar);
-}
-static inline
-void raw_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE(Rbyte, RAW_RO, raw_equal_missing_scalar);
-}
+  SEXP out = PROTECT_N(r_new_logical(size), &nprot);
+  int* p_out = LOGICAL(out);
 
-#undef VEC_DETECT_COMPLETE
-
-#define VEC_DETECT_COMPLETE_BARRIER(SCALAR_EQUAL_MISSING) { \
-  for (R_len_t i = 0; i < size; ++i) {                      \
-    p_out[i] = !SCALAR_EQUAL_MISSING(x, i);                 \
-  }                                                         \
-}
-
-static inline
-void list_detect_complete(SEXP x, R_len_t size, int* p_out) {
-  VEC_DETECT_COMPLETE_BARRIER(list_equal_missing_scalar);
-}
-
-#undef VEC_DETECT_COMPLETE_BARRIER
-
-// -----------------------------------------------------------------------------
-
-static inline void vec_detect_complete_col(SEXP x, R_len_t size, int* p_out);
-
-static
-void df_detect_complete(SEXP x, R_len_t size, int* p_out) {
   // Initialize assuming fully complete
   for (R_len_t i = 0; i < size; ++i) {
     p_out[i] = 1;
   }
 
-  int n_col = Rf_length(x);
+  struct df_short_circuit_info info = new_df_short_circuit_info(size, true);
+  PROTECT_DF_SHORT_CIRCUIT_INFO(&info, &nprot);
+
   const SEXP* p_x = VECTOR_PTR_RO(x);
 
-  for (R_len_t i = 0; i < n_col; ++i) {
-    vec_detect_complete_col(p_x[i], size, p_out);
+  for (r_ssize i = 0; i < n_cols; ++i) {
+    vec_detect_complete_col(p_x[i], size, p_out, &info);
   }
+
+  UNPROTECT(nprot);
+  return out;
 }
+
+// -----------------------------------------------------------------------------
 
 static inline void lgl_detect_complete_col(SEXP x, R_len_t size, int* p_out);
 static inline void int_detect_complete_col(SEXP x, R_len_t size, int* p_out);
@@ -156,23 +94,29 @@ static inline void cpl_detect_complete_col(SEXP x, R_len_t size, int* p_out);
 static inline void chr_detect_complete_col(SEXP x, R_len_t size, int* p_out);
 static inline void raw_detect_complete_col(SEXP x, R_len_t size, int* p_out);
 static inline void list_detect_complete_col(SEXP x, R_len_t size, int* p_out);
+static void df_detect_complete_col(SEXP x, R_len_t size, int* p_out, struct df_short_circuit_info* p_info);
 
 static inline
-void vec_detect_complete_col(SEXP x, R_len_t size, int* p_out) {
-  switch (vec_proxy_typeof(x)) {
-  case vctrs_type_logical: lgl_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_integer: int_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_double: dbl_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_complex: cpl_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_character: chr_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_raw: raw_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_list: list_detect_complete_col(x, size, p_out); break;
-  case vctrs_type_dataframe: stop_internal("vec_detect_complete_col", "Data frame columns should have been handled already.");
+void vec_detect_complete_col(SEXP x, R_len_t size, int* p_out, struct df_short_circuit_info* p_info) {
+  SEXP proxy = PROTECT(vec_proxy_equal(x));
+
+  switch (vec_proxy_typeof(proxy)) {
+  case vctrs_type_logical: lgl_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_integer: int_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_double: dbl_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_complex: cpl_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_character: chr_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_raw: raw_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_list: list_detect_complete_col(proxy, size, p_out); break;
+  case vctrs_type_dataframe: df_detect_complete_col(proxy, size, p_out, p_info); break;
   case vctrs_type_scalar: stop_internal("vec_detect_complete_col", "Can't detect missing values in scalars.");
-  default: stop_unimplemented_vctrs_type("vec_detect_complete_col", vec_proxy_typeof(x));
+  default: stop_unimplemented_vctrs_type("vec_detect_complete_col", vec_proxy_typeof(proxy));
   }
+
+  UNPROTECT(1);
 }
 
+// -----------------------------------------------------------------------------
 
 /*
  * Avoid the temptation to add an extra if branch at the start of the for
@@ -198,7 +142,7 @@ void vec_detect_complete_col(SEXP x, R_len_t size, int* p_out) {
  * cols <- c(list(first), cols)
  * names(cols) <- paste0("a", 1:length(cols))
  * df <- new_data_frame(cols)
- * bench::mark(vec_detect_complete(df), complete.cases(df))
+ * bench::mark(df_detect_complete(df), complete.cases(df))
  * ```
  */
 
@@ -255,3 +199,161 @@ void list_detect_complete_col(SEXP x, R_len_t size, int* p_out) {
 }
 
 #undef VEC_DETECT_COMPLETE_COL_BARRIER
+
+// -----------------------------------------------------------------------------
+
+static void vec_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+
+/*
+ * This data frame path occurs when a vector that isn't a data frame has a
+ * data frame proxy, like a rcrd. In these cases, we consider a "complete"
+ * row to be any row that has at least one non-missing value.
+ */
+
+static
+void df_detect_complete_col(SEXP x, R_len_t size, int* p_out, struct df_short_circuit_info* p_info) {
+  r_ssize n_cols = r_length(x);
+  const SEXP* p_x = VECTOR_PTR_RO(x);
+
+  // Ensure lazy memory is initialized
+  init_lazy_df_short_circuit_info(p_info);
+
+  // Reset between each df-col
+  p_info->remaining = size;
+
+  // If `p_out[i] == true`, we don't yet know the row value since
+  // `true` was the default value. Otherwise we already know it to be an
+  // incomplete row and can skip.
+  for (r_ssize i = 0; i < size; ++i) {
+    const bool known = !p_out[i];
+    p_info->p_row_known[i] = known;
+    p_info->remaining -= known;
+  }
+
+  for (r_ssize i = 0; i < n_cols; ++i) {
+    SEXP col = p_x[i];
+
+    vec_detect_any_non_missing_col(col, size, p_info);
+
+    if (p_info->remaining == 0) {
+      break;
+    }
+  }
+
+  // If we didn't know the row value before (`p_out[i] == true`),
+  // and we still don't (`p_row_known[i] == false`),
+  // then the entire row of this df-col is missing,
+  // which we consider incomplete.
+  for (r_ssize i = 0; i < size; ++i) {
+    if (p_out[i] && !p_info->p_row_known[i]) {
+      p_out[i] = 0;
+    }
+  }
+}
+
+static inline void lgl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void int_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void dbl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void cpl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void chr_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void raw_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+static inline void list_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info);
+
+static
+void vec_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  switch (vec_proxy_typeof(x)) {
+  case vctrs_type_logical: lgl_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_integer: int_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_double: dbl_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_complex: cpl_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_character: chr_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_raw: raw_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_list: list_detect_any_non_missing_col(x, size, p_info); break;
+  case vctrs_type_dataframe: stop_internal("vec_detect_any_non_missing_col", "Data frame columns should have been handled already.");
+  case vctrs_type_scalar: stop_internal("vec_detect_any_non_missing_col", "Can't detect missing values in scalars.");
+  default: stop_unimplemented_vctrs_type("vec_detect_any_non_missing_col", vec_proxy_typeof(x));
+  }
+}
+
+#define VEC_DETECT_ANY_NON_MISSING(CTYPE, CONST_DEREF, SCALAR_EQUAL_MISSING) { \
+  const CTYPE* p_x = CONST_DEREF(x);                                           \
+                                                                               \
+  for (R_len_t i = 0; i < size; ++i) {                                         \
+    /* Known to be incomplete row */                                           \
+    if (p_info->p_row_known[i]) {                                              \
+      continue;                                                                \
+    }                                                                          \
+                                                                               \
+    const CTYPE elt = p_x[i];                                                  \
+                                                                               \
+    if (SCALAR_EQUAL_MISSING(elt)) {                                           \
+      continue;                                                                \
+    }                                                                          \
+                                                                               \
+    /* At least one non-missing value exists */                                \
+    p_info->p_row_known[i] = true;                                             \
+    --p_info->remaining;                                                       \
+                                                                               \
+    /* All rows have at least one non-missing value */                         \
+    if (p_info->remaining == 0) {                                              \
+      break;                                                                   \
+    }                                                                          \
+  }                                                                            \
+}
+
+static inline
+void lgl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(int, LOGICAL_RO, lgl_equal_missing_scalar);
+}
+static inline
+void int_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(int, INTEGER_RO, int_equal_missing_scalar);
+}
+static inline
+void dbl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(double, REAL_RO, dbl_equal_missing_scalar);
+}
+static inline
+void cpl_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(Rcomplex, COMPLEX_RO, cpl_equal_missing_scalar);
+}
+static inline
+void chr_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(SEXP, STRING_PTR_RO, chr_equal_missing_scalar);
+}
+static inline
+void raw_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING(Rbyte, RAW_RO, raw_equal_missing_scalar);
+}
+
+#undef VEC_DETECT_ANY_NON_MISSING
+
+
+#define VEC_DETECT_ANY_NON_MISSING_BARRIER(SCALAR_EQUAL_MISSING) { \
+  for (R_len_t i = 0; i < size; ++i) {                             \
+    /* Known to be incomplete row */                               \
+    if (p_info->p_row_known[i]) {                                  \
+      continue;                                                    \
+    }                                                              \
+                                                                   \
+    if (SCALAR_EQUAL_MISSING(x, i)) {                              \
+      continue;                                                    \
+    }                                                              \
+                                                                   \
+    /* At least one non-missing value exists */                    \
+    p_info->p_row_known[i] = true;                                 \
+    --p_info->remaining;                                           \
+                                                                   \
+    /* All rows have at least one non-missing value */             \
+    if (p_info->remaining == 0) {                                  \
+      break;                                                       \
+    }                                                              \
+  }                                                                \
+}
+
+static inline
+void list_detect_any_non_missing_col(SEXP x, R_len_t size, struct df_short_circuit_info* p_info) {
+  VEC_DETECT_ANY_NON_MISSING_BARRIER(list_equal_missing_scalar);
+}
+
+#undef VEC_DETECT_ANY_NON_MISSING_BARRIER
